Remove some mocking from `Watir::Rails` spec
Use `#run` instad of `#join(0.1)` during server thread startup, it's more
explicit and raises if server thread dies so we don't wait for timeout

Also handle `port: 0` as a random port and handle server thread exceptions

Default Net::HTTP timeouts are 60 seconds, so we're makeing a single request
and waiting for it until boot timeout, which is not ideal. Wait for 1 second.